Fix RoleTurnFace and RolePathTo (#33)
* fix: RoleTurnFace: @degree should be i32, not f32


* fix: RolePathTo: support @run = 2: walk backwards


---------

+++ yaobow/shared/src/scripting/sce/vm.rs
@@ -361,7 +361,7 @@ impl SceProcContext {
             }
             24 => {
                 // RoleTurnFace
-                command!(self, SceCommandRoleTurnFace, role_id: i32, degree: f32)
+                command!(self, SceCommandRoleTurnFace, role_id: i32, degree: i32)
             }
             25 => {
                 // TeamOpen
